requirements.txt
Flask
faiss-cpu
numpy
pypdf
sentence-transformers
llama-cpp-python
langchain
